// src/components/ui/alert.d.ts
import * as React from "react";
import { type VariantProps } from "class-variance-authority";
declare const Alert: React.ForwardRefExoticComponent<
  React.HTMLAttributes<HTMLDivElement> &
    VariantProps<
      (
        props?: {
          variant?: "default" | "info" | "warning" | "destructive";
        } & import("class-variance-authority/dist/types").ClassProp,
      ) => string
    > &
    React.RefAttributes<HTMLDivElement>
>;
declare const AlertTitle: React.ForwardRefExoticComponent<
  React.HTMLAttributes<HTMLHeadingElement> &
    React.RefAttributes<HTMLParagraphElement>
>;
declare const AlertDescription: React.ForwardRefExoticComponent<
  React.HTMLAttributes<HTMLParagraphElement> &
    React.RefAttributes<HTMLParagraphElement>
>;
export { Alert, AlertTitle, AlertDescription };